pyproject.toml
[tool.poetry]
name = "song-do"
version = "0.1.0"
description = ""
authors = ["dearsecret <dearsecret1217@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
python-dateutil = "^2.8.2"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
